server/dm_api/requirements.txt
alembic==1.4.3
aniso8601==8.1.0
appdirs==1.4.4
appnope==0.1.2
argcomplete==1.12.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
boto3==1.16.52
botocore==1.19.52
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
coverage==5.3.1
databricks-cli==0.14.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
docker==4.4.1
dotmap==1.3.0
durationpy==0.5
elastic-apm==5.10.0
elasticsearch==7.10.0
entrypoints==0.3
faiss-cpu==1.6.3
farm==0.5.0
farm-haystack @ git+https://github.com/deepset-ai/haystack.git@e9b5439b00690a92b9e385cae341d47a803f20d5
fastapi==0.63.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flask-restplus==0.13.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
gunicorn==20.0.4
h11==0.12.0
h5py==3.1.0
hjson==3.0.2
httptools==0.1.1
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.4.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupytext==1.9.1
kappa==0.6.0
Keras==2.4.3
langdetect==1.0.8
lxml==4.6.2
Mako==1.1.3
markdown-it-py==0.6.1
MarkupSafe==1.1.1
mdit-py-plugins==0.2.4
mistune==0.8.4
mlflow==1.0.0
more-itertools==8.6.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.1.6
numpy==1.19.5
packaging==20.8
pandas
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==5.5.0
placebo==0.9.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
pydantic==1.7.3
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
python-multipart==0.0.5
python-slugify==4.0.1
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
querystring-parser==1.2.4
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.24.0
scipy
Send2Trash==1.5.0
sentencepiece==0.1.95
seqeval==0.0.12
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smmap==3.0.4
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
starlette==0.13.6
tabulate==0.8.7
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tika==1.24
tokenizers==0.8.1rc2
toml==0.10.2
torch==1.6.0
tornado==6.1
tox==3.21.0
tqdm==4.56.0
traitlets
transformers==3.3.1
troposphere==2.6.3
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
virtualenv==20.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.4.0
